--- src/main/java/com/study/petory/domain/ownerBoard/dto/response/OwnerBoardCommentUpdateResponseDto.java
@@ -1,29 +1,0 @@
-package com.study.petory.domain.ownerBoard.dto.response;
-
-import java.time.LocalDateTime;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.study.petory.domain.ownerBoard.entity.OwnerBoardComment;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class OwnerBoardCommentUpdateResponseDto {
-
-	private final Long id;
-	private final String content;
-	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-	private final LocalDateTime updatedAt;
-	private final Long writerId;
-
-	public static OwnerBoardCommentUpdateResponseDto from(OwnerBoardComment comment) {
-		return new OwnerBoardCommentUpdateResponseDto(
-			comment.getId(),
-			comment.getContent(),
-			comment.getUpdatedAt(),
-			comment.getUser().getId()
-		);
-	}
-}
