--- src/main/java/com/study/petory/domain/ownerBoard/repository/OwnerBoardImageRepository.java
@@ -1,8 +1,0 @@
-package com.study.petory.domain.ownerBoard.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.study.petory.domain.ownerBoard.entity.OwnerBoardImage;
-
-public interface OwnerBoardImageRepository extends JpaRepository<OwnerBoardImage, Long> {
-}
