--- src/main/java/com/study/petory/domain/ownerBoard/repository/OwnerBoardQueryRepository.java
@@ -1,19 +1,0 @@
-package com.study.petory.domain.ownerBoard.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-import com.study.petory.domain.ownerBoard.dto.response.OwnerBoardGetAllResponseDto;
-import com.study.petory.domain.ownerBoard.entity.OwnerBoard;
-
-public interface OwnerBoardQueryRepository {
-
-	Page<OwnerBoardGetAllResponseDto> findAllWithFirstImageAndTitleOptional(String title, Pageable pageable);
-
-	Optional<OwnerBoard> findByIdWithImages(Long boardId);
-
-	Optional<OwnerBoard> findByIdIncludingDeleted(Long boardId);
-
-}
