+++ src/main/java/com/study/petory/domain/ownerboard/dto/response/OwnerBoardGetAllResponseDto.java
@@ -1,0 +1,18 @@
+package com.study.petory.domain.ownerboard.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class OwnerBoardGetAllResponseDto {
+
+	private final Long id;
+
+	private final String title;
+
+	private final String content;
+
+	private final String imageUrl;
+
+}
