+++ src/main/java/com/study/petory/domain/ownerboard/repository/OwnerBoardImageRepository.java
@@ -1,0 +1,8 @@
+package com.study.petory.domain.ownerboard.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.study.petory.domain.ownerboard.entity.OwnerBoardImage;
+
+public interface OwnerBoardImageRepository extends JpaRepository<OwnerBoardImage, Long> {
+}
